:sparkles: (you icon): Move config and context into icon
