+++ packages/components/material-you/icon/src/Icon.svelte
@@ -15,10 +15,82 @@ child of the component.
 ```
 -->
 
+<script lang="ts" context="module">
+  import { getContext, setContext } from "svelte";
+
+  export type IconStyle = "outlined" | "rounded" | "sharp" | `custom__${string}`;
+
+  export type IconFill = boolean | 0 | 1 | `${0 | 1}`;
+
+  export type IconSize = 20 | 24 | 40 | 48 | `${20 | 24 | 40 | 48}`;
+
+  export type IconWeight =
+    | 100
+    | 200
+    | 300
+    | 400
+    | 500
+    | 600
+    | 700
+    | `${100 | 200 | 300 | 400 | 500 | 600 | 700}`;
+
+  export type IconGrade = number | `${number}`;
+
+  export type IconContext =
+    | {
+        style: IconStyle;
+        fill: IconFill;
+        weight: IconWeight;
+        size: IconSize;
+        grade: IconGrade;
+      }
+    | undefined;
+
+  export const defaultIconContext = {
+    style: "outlined",
+    fill: false,
+    weight: 400,
+    size: 24,
+    grade: 0,
+  } satisfies NonNullable<IconContext>;
+
+  const contextKey = Symbol("iconContext");
+
+  export const iconContext = {
+    get: () => getContext<IconContext>(contextKey) ?? defaultIconContext,
+    set: (context: IconContext): void =>
+      void setContext(contextKey, { ...iconContext.get(), ...context }),
+  };
+</script>
+
 <script lang="ts">
   import type { MaterialSymbols } from "./symbols.js";
-  import { iconContext, type IconStyle } from "./context.js";
 
+  /**
+   * Create this component in "config mode", which sets default configurations for other `Icon`
+   * components.
+   *
+   * > **Note:** `alt` and `name` do nothing in this mode
+   *
+   * By passing in props to this component while in config mode, any icons which are children of
+   * this component will inherit those props. The mechanism powering this is
+   * [Svelte Context](https://svelte.dev/docs#run-time-svelte-setcontext), therefore it is
+   * recommended to only use constant values as props to this component while in config mode.
+   *
+   * @example
+   *
+   * ```svelte
+   * <Icon configMode fill weight="500" size="40" grade="200">
+   *   // Filled, weight 500, size 40, grade 200
+   *   <Icon/>
+   *   // Not filled, weight 500, size 40, grade 200
+   *   <Icon fill={false} />
+   *   // Filled, weight 200, size 24, grade 200
+   *   <Icon size="24" weight={200} />
+   * </Icon>
+   * ```
+   */
+  export let configMode = false;
   /**
    * The symbol text to use.
    *
@@ -51,7 +123,7 @@ child of the component.
    * For custom icon packages, you can set this to `custom__<classname>`, where `<classname>` is the
    * name of the CSS class for your custom icon package.
    *
-   * Defaults to the value from the nearest `IconConfig` component, or `"outlined"` if none is found.
+   * Defaults to the value from the nearest `configMode` parent, or `"outlined"` if none is found.
    */
   export let style = ctx.style;
 
@@ -60,30 +132,32 @@ child of the component.
   /**
    * Whether the Icon should be filled or not.
    *
-   * Defaults to the value from the nearest `IconConfig` component, or `false` if none is found.
+   * Defaults to the value from the nearest `configMode` parent, or `false` if none is found.
    */
   export let fill = ctx.fill;
 
   /**
    * The weight of the Icon, from `100` to `900` in increments of `100`.
    *
-   * Defaults to the value from the nearest `IconConfig` component, or `400` if none is found.
+   * Defaults to the value from the nearest `configMode` parent, or `400` if none is found.
    */
   export let weight = ctx.weight;
 
   /**
    * The size of the Icon in pixels, either `20`, `24`, `40`, or `48`.
    *
-   * Defaults to the value from the nearest `IconConfig` component, or `24` if none is found.
+   * Defaults to the value from the nearest `configMode` parent, or `24` if none is found.
    */
   export let size = ctx.size;
 
   /**
    * The grade of the Icon, from `-25` to `200`.
    *
-   * Defaults to the value from the nearest `IconConfig` component, or `0` if none is found.
+   * Defaults to the value from the nearest `configMode` parent, or `0` if none is found.
    */
   export let grade = ctx.grade;
+
+  $: if (configMode) iconContext.set({ style, fill, weight, size, grade });
 
   const getClass = (style: IconStyle) => {
     if (style.startsWith("custom__")) return style.slice(8);
@@ -91,28 +165,32 @@ child of the component.
   };
 </script>
 
-<span
-  class={getClass(style)}
-  style="font-variation-settings: 'FILL' {+fill}, 'wght' {weight}, 'GRAD' {grade}, 'opsz' {size}"
->
-  <span class="sr-only">{alt}</span>
-  <span aria-hidden="true">
-    <slot>
-      {name ||
-        console.error(
-          "No icon name provided.",
-          `props: ${JSON.stringify({
-            name,
-            style,
-            fill,
-            weight,
-            size,
-            grade,
-          })}`,
-        )}
-    </slot>
+{#if !configMode}
+  <span
+    class={getClass(style)}
+    style="font-variation-settings: 'FILL' {+fill}, 'wght' {weight}, 'GRAD' {grade}, 'opsz' {size}"
+  >
+    <span class="sr-only">{alt}</span>
+    <span aria-hidden="true">
+      <slot>
+        {name ||
+          console.error(
+            "No icon name provided.",
+            `props: ${JSON.stringify({
+              name,
+              style,
+              fill,
+              weight,
+              size,
+              grade,
+            })}`,
+          )}
+      </slot>
+    </span>
   </span>
-</span>
+{:else}
+  <slot />
+{/if}
 
 <style lang="scss">
   .sr-only {
